backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
django-filter==23.2
python-decouple==3.8
psycopg2-binary==2.9.6
Pillow==9.5.0
gunicorn==20.1.0
whitenoise==6.4.0 